Combining Source utilities for refactor to LarkSourceManager
