+++ src/Service/Utility/StatusResolver.php
@@ -8,7 +8,6 @@ use Drupal\lark\Model\ExportableStatus;
 use Drupal\lark\Model\ExportableInterface;
 use Drupal\lark\Model\ExportArray;
 use Drupal\lark\Model\LarkSettings;
-use Drupal\lark\Service\ImporterInterface;
 
 /**
  * Used to determine the status of an exportable entity and prepare exports for
@@ -19,16 +18,13 @@ class StatusResolver {
   /**
    * StatusResolver constructor.
    *
-   * @param \Drupal\lark\Service\Utility\SourceResolver $sourceResolver
+   * @param \Drupal\lark\Service\Utility\SourceUtility $sourceUtility
    *   The source resolver.
-   * @param \Drupal\lark\Service\ImporterInterface $importer
-   *   The importer.
    * @param \Drupal\lark\Model\LarkSettings $larkSettings
    *   The lark settings.
    */
   public function __construct(
-    protected SourceResolver $sourceResolver,
-    protected ImporterInterface $importer,
+    protected SourceUtility $sourceUtility,
     protected LarkSettings $larkSettings,
   ) {}
 
@@ -47,7 +43,7 @@ class StatusResolver {
    */
   public function resolveStatus(ExportableInterface $exportable, ?ExportArray $sourceExportArray = NULL): ExportableStatus {
     $entity = $exportable->entity();
-    $source = $this->sourceResolver->resolveSource($exportable);
+    $source = $this->sourceUtility->resolveSource($exportable);
 
     if (!$source) {
       return ExportableStatus::NotExported;
@@ -116,7 +112,7 @@ class StatusResolver {
    * @return array
    *   The processed export array.
    */
-  public function processExportArrayForComparison(array $array): array {
+  protected function processExportArrayForComparison(array $array): array {
     $this->deepUnsetAll($array, $this->larkSettings->ignoredComparisonKeysArray());
     return $array;
   }
